Solve 2020/10
For part b, I realized quickly that I could split into groups where the
difference was 3 and multiply the valid combinations of the subgroups
together. The larger challenge was figuring out how to determine how
many valid combinations there were in the subgroups.

I chose to memoize the valid_chains method realizing that it was likely
to be called multiple times with the same inputs, and it seemed like
an expensive function.