Solve 2020/11
I ran into some oversights during the implementation of this puzzle, but
it's basically my original design. I needed a deep copy because using a
shallow copy made the new row and the old row point to the exact same
object.

Using a list, negative indexes take from the back of the list, so they
would not raise an IndexError, so I had to add my own check for negative
indexes.