Solve 2020/20
The first half of this problem was "simple". There's no need to put
together the tile map, because all that is needed is the corners.

The second half took me a lot of time. Early on, I realized that the
easiest way to deal with the final image was to use a larger version of
the same `Tile` object, so I made sure to keep that as flexible as
possible.

The final search ended up being easier when I figured out that I could
use a series of regular expressions to find the monsters.