Solve 2020/22
Another thought I had was to not create distinct player and card
variables - using lists or tuples, but I don't think that would be very
readable.

Card draws would be a comprehension `(x.popleft() for x in players)`,
but there'd still be conditionals for who wins to extend the deck
appropriately (player2's deck needs to get extended in reverse order).