Solve 2020/24
It took me a while to comprehend how to work with a hexagonal grid in
simple 2 dimensional space, but a little googling, really helped out -
pointing out that the grid is really just offset on every row.

Finding the concept of the skewed axis coordinate system really made
this much more managable, although when plotting on paper, it's
important to understand that the Y coordinates are "diagonal". This
coordinate system works with vector sums - a huge advantage for this
puzzle.