Solve 2015/11
I first tried to use a regular expression to detect how many z's were at
the end of the string, until I realized that a simple recursive call
could handle incrementing the substring.

The search for pairs of strings would have been more complicated to do
without the regular expression. The trick was to use a back-referenced
capture group to be able to detect that the second character was the
same as the first. Then `len(findall())` made the validation straight
forward.