Solve 2015/15
This took a while to come up with a good data structure to work with.
Eventually, I realized that transposing the data would make things
simple - using map() to handle the ingredient multiplication.

The assumptions made in this implementation are that the properties are
in the same order for each ingredient, and that calories are always
specified last.

It is somewhat slow - there's a large number of loops that could be
considered pointless, but the alternative nested loop would have been
really ugly. I suspect that the time is really in computing the value of
each viable recipe anyway (there's ~157,000 recipes to compute, and
~756,000 with too many ingredients).