Solve 2015/20
This was rather difficult. After a large amount of analysis, I finally
realized that this is a sum of all factors. Eventually, I remembered
that I had implemented most of the routines in Project Euler.

On the second part, I kept trying to figure out a mathematical means to
determine the result, but eventually realized that trying to find the
sum of all common factors. I kept thinking that what I was doing was
"wrong" because it was taking too long to calculate the final result.
I finally realized that where I was wrong was trying to sum the common
factors for every house.