Solve 2015/21
Itertools really simplified this problem.

The first simplification was realizing that I could use a single
combination for each equipment type, simulating no armor with 0 cost
and no assistance. I extended that concept to rings (2 rings that cost
0 and provided no assistance). Without this, I would have needed more
loops to pick 0 or 1 armor, and 0-2 rings.

The second simplification was using product to provide all of the
different ways the 1 weapon, 1 armor, and 2 rings could be put together.
Since this product pruduced tuples of tuples (for weapon, a 1-tuple of
weapon stats), I simplified the equipment with chain.

In the end, every cycle in the loop, the player is equipped with exactly
4 items (the armor and rings could be worthless).