Solve 2015/22
This was a rather interesting case that helped with my class building.
I'd still like to further simplify how the spells work. Treating missile
and drain as a 1 turn effect slows it down by about 10%. What I'm really
considering next is a decorator for the spell effect that would
automatically register the spell in Player._spells.

```
@spell(cost=229, duration=5)
def _recharge(self):
    pass
```