Solve 2016/11
For the first part, it was easy to use an itertools.product() to step
through the field, but it ended up not being feasible to use that type
of logic for the second part. In particular, range() does not accept a
step of 0. Options would be to use zip() where the arguments would be
either a range(), or an itertools.count(), or itertools.zip_longest()
with a fillvalue, but those would really be complicated to read.