Solve 2021/14 part b
The key was to realize that we only really needed to keep track of the
number of occurances of each pair. During each expansion, the pairs
completely mutate, which is why we create a new counter in every
expansion.

The final step is counting the first character of every pair, plus the
last character of the original template.