Solve 2021/15
This is currently way to slow for part b (~ 1 hour), but it does provide
the correct answer. I'm surprised that my optimization of removing the
visited set and deleting the item from the grid don't really seem to
help. This must mean that finding the next shortest path to try is the
slow part (all current paths - the ones that are no longer testable).