Solve 2021/16
I had actually started to write an EBNF grammar for lark, but then
realized that the concept of the operator length was not possible to do.
I had also started creating a "stream parser" (reading 1 byte at a
time), but that also got rather complex. In the end, this offset
management solution worked quite well.