Solve 2016/22
The puzzle description gave a pretty big clue regarding how to solve
this - only "move" the 1 empty node around. That meant that it was a
breadth-first search, but the location of the goal data was the only
determining factor for the end state. It doesn't matter if the empty
node is below or to the right of (0, 0).