Solve 2021/19
I had to learn how to use matrix calculations to simplify the concept of
rotations. I therefore decided to add in numpy, which provides matrix
calculations - there's no need to re-implement that complex logic.

I've also learned that numpy also handles a lot of calculations rather
automatically. For example, coordinate adding and subtraction can be
handled very easily, although numpy arrays are not hashable, meaning
they cannot be placed in a `dict` or `set`, so some conversions to
`tuple` are warranted.