Solve 2021/23
This is not really that efficient, unfortunately. It takes about 10
minutes to work through all of the possibilities. One possible
optimization would be to convert seen to a dict that maintains the best
cost to get to that state. This will require changing the state to no
longer include the cost itself.

Another improvement could be to add a room-to-room movement function. If
room-to-room moves are available, make those and return. If not, check
for hall-to-room moves, then for room-to-hall moves.

I had to use pop() (depth-first search) instead of popleft()
(breadth-first search) due to the size of the queue. Perhaps by ordering
the moves, we can try "low-cost" moves that are likely to result in a
non-completable state first.