Solve 2021/24
This was a rather difficult one to try to figure out. With 14 digits, I
thought it might encode YYYY-MM-DD-HH:MM:SS (which i was expecting to
end up as 2021-12-25-00:00:00 or something similar), but it doesn't seem
to be.

Eventually, I noticed that the program was a repeated sequence of 18
commands with different constants. Each "function" basically compared
the current digit to the previous digit. The variable x stored the
comparsion where if it was 0, the value of z would be reduced. This lead
me to the solution of working each digit, and only keeping track of
those results where x == 0 (meaning that z would be reduced).  Since
some of the comparisons would always end in x==1, I needed to try all of
those with the next digit. Doing this reduced the total number of
permutations significantly (from 9^14 to ~ 3,000,000).