Solve 2021/22 part b
There must be a more efficient way to do this - at the moment, this
takes about 7 minutes to run. However, it does solve the problem.

The only thing I can think of would be to somehow store the current
"volume" of a cuboid, reducing the volume by the amount of overlap.