Solve 2016/24
This algorithm ended up being faster than expected. I was concerned that
the algorithm was going to take a long time to find all of the possible
edges. From there, the next concern would be the number of permutations,
which in this case is only 5040, so not that big of a deal.

The thing that really made the path-finding algorithm efficient was by
determining the end as when it found paths to all of the other
destinations (instead of searching for the fastest way between A->B,
then A->C, etc).