Solve 2022/11
Obviously, `eval` is frowned upon for security reasons, but the
restrictive regex mitigates that problem. The regex only allows for an
operator surrounded by two operands, both of which are restricted to
either the identifier `old`, or a digit.

The second part required the use of lowest-common-multiple. Without it,
the script can only go up to about 100 rounds before it becomes
untenable.