Solve 2023/03 part b
After proving that numbers only have a single symbol neighboring them, I
refactored use the search for symbols as the main outer loop. That made
it easy to search just for `*` symbols instead of all symbols and
determine which `*`s were actually gears.

An improvement would be to create a Part class, that only tracks the key
elements that matter.