Solve 2023/05
The first part was rather simple, although my initial solution doesn't
really exist in here. That was basically looping over the maps and
running `.get()`.

For the second part, I recognized that I really needed to look up the
start of the ranges, but it took me a while to get there. I briefly
tried to create a single map by splitting the individual rows, but
couldn't really come up with the logic. Finally I realized that the only
ranges I needed to worry about splitting were the lookup ranges, because
in the end, I only need to compare the start of each range.