Solve 2023/06 faster
This only determines the number of worse options from either end,
eliminating the middle section.

A better solution yet would be a binary search pattern. In this
solution, if the previous best were close to the actual best, it would
take about the same amount of time as calculating the full range.