Solve 2023/14
Numpy really made this rather easy, since slices and transformations are
views into the same ndarray. The ability to do a multi-dimensional slice
([:, ::-1]) was also really helpful.

Part of the reason I chose numpy was because I expected part 2 would
include tilting other directions, but it was still really convenient
just for tilting north, since that really needed to operate on columns.