Solve 2023/24 part a
I'm not sure how to deal with the second part - it would seem that I
need to find a tansversal line that intersects all of the lines, but so
far, i've not found a way to do so.

Maybe there's a different, linear algebra way to accomplish this, but
unfortunately, I can't really remember any linear algebra details.